pyproject.toml
[tool.poetry]
name = "pulsar-neuron"
version = "0.1.0"
description = "Deterministic trading lib + agentic brains for NIFTY/BANKNIFTY."
authors = ["Abhay <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "pulsar_neuron", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
pytz = "^2024.1"
pendulum = "^3.0.0"
numpy = "^2.1.0"
pandas = "^2.2.2"
loguru = "^0.7.2"
tenacity = "^9.0.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
python-telegram-bot = "^21.5"
prometheus-client = "^0.20.0"
pyyaml = "^6.0.2"
boto3 = "^1.35.20"
botocore = "^1.35.20"

[tool.poetry.group.cloud.dependencies]
google-cloud-secret-manager = "^2.20.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.1"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
pre-commit = "^3.8.0"
types-requests = "^2.32.0.20240914"
types-pyyaml = "^6.0.12"
types-boto3 = "^1.0.2"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
py_version = 312
line_length = 100

[tool.pytest.ini_options]
addopts = "-q --disable-warnings"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = true
